+++ src/types/index.ts
@@ -3,7 +3,7 @@ export type { IBounds } from './IBounds.js';
 export type { ICadence } from './ICadence.js';
 export type { IDateLiteral } from './IDateLiteral.js';
 export type { IEvalOptions } from './IEvalOptions.js';
-export type { IDtreIR, IExpressionIR } from './IExpressionIR.js';
+export type { IDTRExpIR, IExpressionIR } from './IExpressionIR.js';
 export type { IFields } from './IFields.js';
 export type { IInterval } from './IInterval.js';
 export type { ISelector, IStride } from './ISelector.js';